Controlando la pagina home para mostrar todos los cursos

+++ database/factories/ReviewFactory.php
@@ -9,7 +9,7 @@ $factory->define(Review::class, function (Faker $faker) {
     return [
         
         'course_id' => \App\Course::all()->random()->id,
-        'rating' => $faker->randomFloat(2, 1, 5),
+        'rating' => $faker->randomInteger(2, 1, 5),
 
     ];
 });
